+++ apps/common/management/commands/services/services/gunicorn.py
@@ -1,0 +1,30 @@
+from ..hands import *
+from .base import BaseService
+
+
+class GunicornService(BaseService):
+
+    def __init__(self):
+        super().__init__(name='gunicorn')
+
+    @property
+    def cmd(self):
+        log_format = '%(h)s %(t)s %(L)ss "%(r)s" %(s)s %(b)s '
+        bind = f'{HTTP_HOST}:{HTTP_PORT}'
+        cmd = [
+            'gunicorn', 'jumpserver.wsgi',
+            '-b', bind,
+            '-k', 'gthread',
+            '--threads', '10',
+            '-w', str(WORKERS),
+            '--max-requests', '4096',
+            '--access-logformat', log_format,
+            '--access-logfile', '-'
+        ]
+        if DEBUG:
+            cmd.append('--reload')
+        return cmd
+
+    @property
+    def cwd(self):
+        return APPS_DIR
